Add sign out button, clean up NavBar
