+++ frontend/src/app/page.tsx
@@ -1,12 +1,10 @@
 import {PrimarySection} from '@/app/components/Section'
 import {PrimaryContainer} from '@/app/components/Container'
 import {Carousel, CarouselSlide} from '@/app/components/Carousel'
-import {MenuButton, MenuContent, SearchField, SiteTitle} from "@/app/layout/NavBar"
-import {SignInModal, SignUpModal} from "@/app/layout/SignUpModal"
-import Link from "next/link"
 import {getRestData, postRestData} from '@/app/utils/fetch'
 import {TagFilterList} from '@/app/page.client'
 import {getSession} from "@/utils/fetchSession";
+import {NavBar} from '@/app/layout/NavBar'
 
 type HomePageProps = {
     searchParams: {
@@ -21,7 +19,7 @@ export default async function HomePage ({searchParams}: HomePageProps) {
     const tagArray = searchParams.tags?.split(',').filter(value => value != '')
     const tags = new Set<string>(tagArray)
     const shopData = tags?.size > 0 ? await postRestData('/apis/shop/getShopsWithTags', JSON.stringify(tagArray)) : await getRestData('/apis/shop')
-    const searchResult = await getRestData(`/apis/shop/search?name=${query}`)
+
     const brewingTags = {
         group: 'Brewing',
         tags: await getRestData('/apis/tag/tagGroup/brewing')
@@ -36,41 +34,7 @@ export default async function HomePage ({searchParams}: HomePageProps) {
     }
     return (
         <>
-            <nav className={'navbar'}>
-                <div className={'dropdown'}>
-                    <MenuButton/>
-                    <ul className={'menu menu-sm dropdown-content mt-3 z-50 p-2 shadow bg-base-100 rounded-box w-32 gap-1'}>
-                        <MenuContent session={session}/>
-                    </ul>
-                </div>
-                <div className={'flex-1'}>
-                    <SiteTitle/>
-                </div>
-                <div className={'flex-none'}>
-                    Search:&nbsp;
-                    <div className={'dropdown'}>
-                        <SearchField initialText={query}>
-                            <div tabIndex={0}>
-                                <ul tabIndex={0}
-                                    className={'dropdown-content z-10 menu grid p-2 shadow bg-base-100 rounded-box sm:w-40 md:w-52 max-h-52 overflow-y-auto gap-4'}>
-                                    {searchResult.length > 0 ? searchResult.map((shop: any) => <Link
-                                            href={`/shop/${shop.shopId}`}>
-                                            <li key={shop.shopId}>{shop.shopName}</li>
-                                        </Link>) :
-                                        <p>No Results</p>}
-                                </ul>
-                            </div>
-                        </SearchField>
-                    </div>
-                    <div className={'navbar-center hidden md:flex'}>
-                        <ul className={'relative flex items-center px-1 gap-4'}>
-                            <MenuContent session={session}/>
-                        </ul>
-                    </div>
-                </div>
-                <SignUpModal/>
-                <SignInModal/>
-            </nav>
+            <NavBar query={query} session={session}/>
             <PrimarySection>
                 <PrimaryContainer autoMargins>
                     <div className={'flex flex-col-reverse sm:flex-col-reverse md:flex-col lg:flex-col'}>
