+++ frontend/src/app/shop/[shopId]/page.tsx
@@ -4,10 +4,8 @@ import {Container} from '@/app/components/Container'
 import {TagToggleList} from '@/app/shop/[shopId]/page.client'
 import {ImageProps, PageProps} from '@/app/types/Props'
 import {getRestData} from "@/app/utils/fetch"
-import {MenuButton, MenuContent, SearchField, SiteTitle} from "@/app/layout/NavBar"
-import Link from "next/link"
-import {SignInModal, SignUpModal} from "@/app/layout/SignUpModal"
 import {getSession} from '@/utils/fetchSession'
+import {NavBar} from '@/app/layout/NavBar'
 
 
 type ShopPageProps = PageProps & {
@@ -21,44 +19,9 @@ export default async function ShopPage({params, searchParams}: ShopPageProps) {
     const photoData = await getRestData(`/apis/photo/photoByShopId/${shopId}`)
     const tagData = await getRestData(`/apis/tag/shopTags/${shopId}`)
     const query = searchParams.q
-    const searchResult = await getRestData(`/apis/shop/search?name=${query}`)
 
     return <>
-        <nav className={'navbar'}>
-            <div className={'dropdown'}>
-                <MenuButton/>
-                <ul className={'menu menu-sm dropdown-content mt-3 z-50 p-2 shadow bg-base-100 rounded-box w-32 gap-1'}>
-                    <MenuContent session={session}/>
-                </ul>
-            </div>
-            <div className={'flex-1'}>
-                <SiteTitle/>
-            </div>
-            <div className={'flex-none'}>
-                Search:&nbsp;
-                <div className={'dropdown'}>
-                    <SearchField initialText={query}>
-                        <div tabIndex={0}>
-                            <ul tabIndex={0}
-                                className={'dropdown-content z-10 menu grid p-2 shadow bg-base-100 rounded-box sm:w-40 md:w-52 max-h-52 overflow-y-auto gap-4'}>
-                                {searchResult.length > 0 ? searchResult.map((shop: any) => <Link
-                                        href={`/shop/${shop.shopId}`}>
-                                        <li key={shop.shopId}>{shop.shopName}</li>
-                                    </Link>) :
-                                    <p>No Results</p>}
-                            </ul>
-                        </div>
-                    </SearchField>
-                </div>
-                <div className={'navbar-center hidden md:flex'}>
-                    <ul className={'relative flex items-center px-1 gap-4'}>
-                        <MenuContent session={session}/>
-                    </ul>
-                </div>
-            </div>
-            <SignUpModal/>
-            <SignInModal/>
-        </nav>
+        <NavBar query={query} session={session}/>
         <PrimarySection>
             <Container autoMargins>
                 <div
